pyproject.toml
[tool.poetry]
name = "e621-dl"
version = "2.0"
description = ""
authors = []
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
rich = "^12.0.1"
requests = "^2.27.1"
typer = "^0.4.0"
e621-temp = "^0.0.2"
click = "<8.1.0"


[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[tool.poetry.scripts]
e6 = 'e621_dl:app'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
